refactor: split /users/me/payments/new into files
